% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOptim.R
\name{DOptim}
\alias{DOptim}
\title{Diagonal Newton optimization for f(d)}
\usage{
DOptim(
  A,
  d0 = NULL,
  tol = 1e-04,
  max_newton_iter = 100,
  max_ls_steps = 15,
  alpha = 0,
  diagonal_Newton = TRUE
)
}
\arguments{
\item{A}{symmetric matrix (p x p); A = R * S}

\item{d0}{initial vector (defaults to ones)}

\item{tol}{convergence tolerance}

\item{max_newton_iter}{max outer Newton iterations}

\item{max_ls_steps}{max line-search iterations}

\item{alpha}{regularization parameter}
}
\value{
list with D, iterations, and final value
}
\description{
Diagonal Newton optimization for f(d)
}
