% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcglassoFast.R
\name{function_to_optimize}
\alias{function_to_optimize}
\title{Function that the `pcglassoFast()` is maximizing}
\usage{
function_to_optimize(R, d, C, lambda, alpha)
}
\description{
In `ROptim()`, this function is maximized with respect to `R`.
In `DOptim()`, this function is maximized with respect to `D`.
}
\details{
\eqn{function\_to\_optimize(R, D) = log(det(R)) + (1-\alpha)log(det(D^2)) - tr(DCDR) - \lambda ||R||_{1,off}}
function, where \eqn{||R||_{1,off}} is only for off-diagonal elements.
}
