% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassoFast.R
\name{glassoFast}
\alias{glassoFast}
\title{Fast graphical LASSO}
\usage{
glassoFast(
  S,
  rho,
  thr = 1e-04,
  maxIt = 10000,
  start = c("cold", "warm"),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE
)
}
\arguments{
\item{S}{Covariance matrix (a p by p symmetric matrix).}

\item{rho}{Regularization parameter (a non-negative value or a p by p matrix).}

\item{thr}{Threshold for convergence. Default is 1e-4.}

\item{maxIt}{Maximum number of iterations. Default is 10,000.}

\item{start}{Type of start: \code{"cold"} or \code{"warm"}.}

\item{w.init}{Optional starting values for the estimated covariance matrix (p x p). Used only for warm starts.}

\item{wi.init}{Optional starting values for the inverse covariance matrix (p x p). Used only for warm starts.}

\item{trace}{Logical. If \code{TRUE}, prints iteration info.}
}
\value{
A list with the following components:
\describe{
  \item{w}{Estimated covariance matrix}
  \item{wi}{Estimated inverse covariance matrix}
  \item{errflag}{Memory allocation error flag: 0 = no error}
  \item{niter}{Number of iterations}
}
}
\description{
A faster alternative to the \code{glasso} function in the \pkg{glasso} package.
This implementation wraps the FORTRAN subroutine by Sustik and Calderhead (2012).
}
\details{
Estimates a sparse inverse covariance matrix using a lasso (L1) penalty, following the approach of Friedman et al. (2008).
}
\examples{
set.seed(100)
p <- 5
x <- matrix(rnorm(p*p), ncol=p)
s <- var(x)
glassoFast(s, rho = 0.1)

}
\references{
Friedman J., Hastie T., Tibshirani R. (2008). Sparse inverse covariance estimation with the graphical lasso. *Biostatistics*, 9, 432–441.\cr
Sustik M.A., Calderhead B. (2012). GLASSOFAST: An efficient GLASSO implementation. *UTCS Technical Report TR-12-29*, 1–3.
}
\seealso{
\code{\link[glasso]{glasso}}
}
\author{
Julien Clavel
}
\keyword{covariance}
\keyword{glasso}
\keyword{likelihood}
\keyword{matrix}
\keyword{penalized}
\keyword{regularization}
