% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{loss.evaluation}
\alias{loss.evaluation}
\title{Loss evaluation}
\usage{
loss.evaluation(Precision.array, Sigma, n, gamma = 0.5)
}
\arguments{
\item{Precision.array}{(p x p x k)}

\item{Sigma}{(p x p) the covariance matrix}

\item{n}{(int)   number of observations}

\item{gamma}{(1 x 1) scaling parameter for \eqn{BIC_{gamma}}}
}
\value{
list (k x 1) $loglik the average loglikelihood
                     $forbenious
                     $n.param
                     $BIC_gamma the average BIC_gamma
}
\description{
computes the loss for the solution of the pcglassoPath, or a list or an array of
matrices
the BIC_gamma is from Extended Bayesian Information Criteria for Gaussian
Graphical Models
}
