% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROptim.R
\name{ROptim}
\alias{ROptim}
\title{Fast optimization for the corelation matrix}
\usage{
ROptim(
  S,
  rho,
  thr = 1e-04,
  maxIt = 10000,
  start = c("cold", "warm"),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE
)
}
\arguments{
\item{S}{Covariance matrix (a p by p symmetric matrix).}

\item{rho}{Regularization parameter (a non-negative value or a p by p matrix).}

\item{thr}{Threshold for convergence. Default is 1e-4.}

\item{maxIt}{Maximum number of iterations. Default is 10,000.}

\item{start}{Type of start: \code{"cold"} or \code{"warm"}.}

\item{w.init}{Optional starting values for the estimated covariance matrix (p x p). Used only for warm starts.}

\item{wi.init}{Optional starting values for the inverse covariance matrix (p x p). Used only for warm starts.}

\item{trace}{Logical. If \code{TRUE}, prints iteration info.}
}
\value{
A list with the following components:
\describe{
  \item{w}{Estimated covariance matrix}
  \item{wi}{Estimated inverse covariance matrix}
  \item{errflag}{Memory allocation error flag: 0 = no error}
  \item{niter}{Number of iterations}
}
}
\description{
Part of the PCGLASSO function.
This implementation is a small modification of the `glassoFast` function from `glassoFast` package.
}
\keyword{covariance}
\keyword{glasso}
\keyword{likelihood}
\keyword{matrix}
\keyword{penalized}
\keyword{regularization}
