% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOptim.R
\name{DOptim}
\alias{DOptim}
\title{using diagonal approximation fo Hessian}
\usage{
DOptim(
  A,
  D0 = NULL,
  tol = 1e-04,
  max.starting.iter = 500,
  max.outer.iter = 100,
  alpha = 0
)
}
\description{
using diagonal approximation fo Hessian
}
