% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrepresentability.R
\name{irrepPCGLASSO}
\alias{irrepPCGLASSO}
\alias{irrepGLASSO}
\title{Compute irrepresentability for GLASSO and PCGLASSO}
\usage{
irrepPCGLASSO(K)

irrepGLASSO(K)
}
\arguments{
\item{K}{True precition matrix}
}
\value{
A non-negative number.
 When smaller than 1, it means the irrepresentability condition holds.
 When bigger than 1, it means the irrepresentability condition is violated.
}
\description{
When the irrepresentability condition holds, then
 for infinite data (\eqn{n = \infty}) there exists a lambda
 that the PCGLASSO (or GLASSO) estimator can recover the true
 underlying structure of the precision matrix.
}
\examples{
set.seed(123)
p <- 7
R.true <- toeplitz(c(1, -0.2, 0, 0, 0, 0, 0))
D.true <- sqrt(rchisq(p, 3))
Q.true <- diag(D.true) \%*\% R.true \%*\% diag(D.true)

irrepPCGLASSO(Q.true) # `0.3999698`

D.true <- sqrt(rchisq(p, 30))
Q.true <- diag(D.true) \%*\% R.true \%*\% diag(D.true)

irrepPCGLASSO(Q.true) # `0.3999698`, the same
set.seed(123)
p <- 7
R.true <- toeplitz(c(1, -0.2, 0, 0, 0, 0, 0))
D.true <- sqrt(rchisq(p, 3))
Q.true <- diag(D.true) \%*\% R.true \%*\% diag(D.true)

irrepGLASSO(Q.true) # `2.588011`

D.true <- sqrt(rchisq(p, 30))
Q.true <- diag(D.true) \%*\% R.true \%*\% diag(D.true)

irrepGLASSO(Q.true) # `0.5570903`, different; depends on `D.true`
}
